persist mint order, add schema versioning
